init.qcom.usb: update handling of persist composition
When PRODUCT_COMPATIBLE_PROPERTY is enabled, the SELinux policy
disallows qti_init_shell from getting/setting the
persist.sys.usb.config property.

Instead, rename the property as persist.vendor.usb.config
and replace the per-target handling to set this property instead
when the property is never set on first bootup. Define a property
trigger that when sys.usb.config is set as ADB-only and
persist.vendor.usb.config is defined to immediately switch
to this default.

Change-Id: I431b974c965de4e9736019254880d1fd130351ef