ANDROID_BUILD_TOP is deprecated
The ANDROID_BUILD_TOP variable is deprecated and thus all paths are
assumed to be relative to the top directory.

However, this is causing problem when some of the paths are constructed
in the Android build system and used in the non-Android makefiles that
are usually executed NOT at the top directory but at a sub directory
using -C option. As a result, the relative path from the Android build
system become invalid in the non-Android makefiles.

In order to overcome the restriction, ANDROID_BUILD_TOP is manually
constructed when invoking the non-Android makefiles by using $$(pwd).
The non-Android makefiles are expected to construct the absolute path
by prepending ANDROID_BUILD_TOP to any path that came from the Android
build system

Change-Id: Ifd2dc5e30af9ec1c4524d98864c7ee6a33f4220a