common: Change permissions or executables launched through init
Some executables are only launched through the init process and
are not supposed to be executed from the shell or from any other
process in general. Setting the permission of these executables
to 744 to prevent processes except those with root UID to execute
these scripts. Keeping the permissions for the root process same
as earlier.

Change-Id: Ib1ee19196de6473465dac4215de71f414c662084