init.msm.usb.configfs.rc: Add support for HID compositions
Add support for below USB compositions to enable HID

• 0x9118 - DIAG + Audio (UAC1.0) + HID
• 0x9119 - DIAG + Audio (UAC1.0) + HID + ADB
• 0x9120 - DIAG + Audio (UAC1.0) + Video (UVC1.5) + HID
• 0x9121 - DIAG + Audio (UAC1.0) + Video (UVC1.5) + HID + ADB
• 0x9122 - DIAG + Audio (UAC2.0) + HID + ADB
• 0x9123 - DIAG + Audio (UAC2.0) + Video (UVC1.5) + HID + ADB

user need to push 'hid_report_desc.bin' to /data/vendor/misc
so that it can be written to HID kernel driver on boot,
which sends this info to host whenever it asks for
report descriptor. this descriptor will have info for
host on how to interpret each control and how to use it.
By having this descriptor as bin and writing to sysfs
gives flexibility to customers to update descriptor or
use thier own descriptor as per thier requirements.

Change-Id: I665792ca6c8b30f3a17893eedb3ddede0320dca8