Build_external_kernelmodule.mk: Only enable DDK per-build
Currently, if any module sets ENABLE_DDK_BUILD, it is set that way
for every modules. This is due to Make's up-front parsing/exapnsion
logig.

Change the variable name to LOCAL_MODULE_DDK_BUILD and add logic to
reset the variable every time its included. Also add a target
dependency to each rule so the enablement is only applied to one rule
at a given time.

Change-Id: Ic9a19e7a299e33d132efd64d61a645e1a35a4c0f