Build_external_kernelmodule: Fix grep search in symvers
Currently, the Makefile will filter the Modules.symvers file based
on the ko name, but does not factor in if the ko is in a subdir.

This caused some Modules.symvers to be empty in the output.

Add a call to notdir to make sure we're searching for just the file
basename (not including the directory).

Change-Id: I54e230b61dc215564507f724441af4598d3ae86f