Init : Decouple the target releated part in the script.
When the device boots up , it only needs to source it's own target script ,
rather than all of them.Given this reason , we do filter in the head to save boot time.
The etr address of all target are not exactly the same, so we move the set_permission into their own script.

Change-Id: I15e6f4ee48c3f0917339b1b71ab7b648acc42fba